move event counting to tools
